requirements.txt
# Core
pandas>=1.5.0
numpy>=1.20.0
requests>=2.28.0
matplotlib>=3.5.0
plotly>=5.0.0
kaleido>=0.2.0
pyarrow>=10.0.0
python-dotenv>=1.0.0

# Web server (React mode)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Web server (Gradio fallback)
gradio>=4.0.0

# LLM providers (install the ones you need)
openai>=1.0.0
google-generativeai>=0.4.0
anthropic>=0.20.0

# Optional
duckduckgo-search>=3.0.0
seaborn>=0.12.0
jupyter>=1.0.0
